requirements.txt
# Production Dependencies
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# CrewAI and AI
crewai>=0.80.0
google-generativeai>=0.3.0
openai>=1.0.0

# Database & ORM
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0  # PostgreSQL async driver
redis>=5.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Background Tasks
celery>=5.3.0

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=23.0.0

# Configuration
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Testing (Development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For testing async endpoints

# Development Tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.6.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Performance
gunicorn>=21.2.0